Disable zone filter, add flag to set zones manually
This the only sane way to fix UnsupportedAvailabilityZoneException,
espcially because it turns out that this is somehow different in every
AWS account (at least that what seems to be the case from the docs.

> You may receive an error that one of the Availability Zones in your
> request does not have sufficient capacity to create an Amazon EKS
> cluster. If this happens, the error output contains the Availability
> Zones that can support a new cluster. Retry creating your cluster
> with at least two subnets that are located in the supported
> Availability Zones for your account.

In the future we may try parsing the error message and retrying
automatically, but it doesn't seem feasible right now as we would
need to improve some of the internals and provide stack update
functionality.