Added MaybeDeleteConfig test and helper functions
Checks to see if MaybeDeleteConfig removes a cluster from the kubeconfig when deleting the cluster if the kubeconfig has the cluster's information and does not change the kubeconfig if it does not.

Also adds functions to supportthe test, such as managing the KUBECONFIG env variable, getting a clusterconfig example, and converting example kubeconfig yaml files into bytes.