Added nodegroup scaling
Initial version of nodegroup scaling has been added. This scales
by modifying the CloudFormation template for the nodegroup. The
modified template is used to create a changeset that is then
executed.

When scaling down/in (i.e. reducing the number of nodes) we rely
solely on the the resulting change to the ASG. This means that
the node(s) that are to be terminated aren't drained and so pods
running on the terminating nodes may cause errors. In the future
we may consider picking the EC2 instances to be terminated and then
drain the nodes and then create a termination policy to ensure those
nodes are killed.

Issue #116

Signed-off-by: Richard Case <richard.case@outlook.com>