Simplify stack output getter
Originally it was assumed this will need to be generic, as the number
of outputs was assumed to increase over time and evolve a lot.
However, as ClusterConfig destination struct has evolved, this was no
longer useful and we needed an intermediary struct.
NodeGroup struct has evolved also and needed intermediary struct too.
The number of outputs turned out to stabilise over time and this kind
complexity is pretty useless now.