Make scaling operations for nodegroups idempotent
CloudFormation throws an error when attempting to submit a changeset to a
stack that contains no actual changes.  This can occur when you attempt to
scale a NodeGroup to the same size that it already is.  To work-around
that, we can detect if you're attempting to scale the NodeGroup to the
existing DesiredCapacity and turn that action into a no-op in order to
make this API call idempotent.

Fixes #430