Fix setup instructions to include dep dependency
Was trying to do a build of the project to test out a currently open PR, and didn't have a preexisting Go setup on my machine.  Looks like you need dep installed to do the `make install-build-deps` step?  

Adding a bit to the documentation to save someone else time in the future