Filter out serviceaccounts that exist in Kubernetes
Due to security considerations default behaviour of
`eksctl create iamserviceaccount` should exclude
updates to the metadata on serviceaccounts that already
exit in Kubernetes. That is in order to prevent a
case where user inadvertedly grats new priveleges to
some pods that are already running.
It's important to note that there may exist pods that
are referencing a serviceaccount which do not exists,
so it's still possible to grant access inadvertedly,
however there is not much that we can do about that
and users should consider an apropriate policy and
RBAC rules to prevent security risks.