No longer default SSH allow to true when SSH Public Keys are defined and SSH allow is set to false.
This change was made to allow people to be able to define a KeyName
without an ingress rule of port 22 open to the VPC being set as a side
effect. This feature is useful in the case where you want to manage
SSH access with your own seperate security group but you still want to
explicitly set the SSH key.

Changed the logic for defaults so that when you define an SSH key and
set SSH to disabled it will no longer set SSH to enabled.

Removed the requirement of having SSH.Allow set to true when defining the
PublicKeyName in the template.

Provided an example where `allow: false` is set with a comment about what the
new behavior is in the ssh-keys example.