Fargate: pkg/fargate.Client#CreateProfile can now wait for creation (+tests)
Waiting is useful as the Fargate API returns the following error
when trying to create a Fargate profile while another still has its
status set to "CREATING":

```
[✖]  failed to create Fargate profile "dev" on EKS cluster "cluster-16": failed to create Fargate profile "dev" in cluster "cluster-16": ResourceInUseException: Cannot create Fargate Profile dev because cluster cluster-16 currently has Fargate profile default in status CREATING
	status code: 409, request id: 88b6c847-47ac-44eb-ad89-0df199d7e247
```

In particular, if multiple profiles are defined in `ClusterConfig`,
then this error is hit. We therefore need to linearise the creation
of all Fargate profiles by waiting each of them to reach an
appropriate status before moving on to the creation of the next one.