Remove support for --wait flag in `create fargateprofile`
When multiple profiles are present in a config file, the second profile fails to create because, by default, we don't wait for the previous profile to become ACTIVE before initiating creation of the second profile. The fact that the AWS Fargate API disallows creation of a profile if another profile is still being created, is an implementation detail of eksctl which users shouldn't be exposed to, or have control over. The `--wait` flag allows a user to break profile creation and hence should not be supported.