Permissions Boundary
A permissions boundary is an advanced IAM feature in which you set the
maximum permissions that an identity-based policy can grant to an IAM
entity; where those entities are either users or roles. When you set
a permissions boundary for an entity, the entity can perform only the
actions that are allowed by both its identity-based policies and its
permissions boundaries.

This is relevant for eksctl, because eksctl is capable of managing,
i.e., creating, updating and deleting, the identity-based resources
required to operate a kubernetes cluster on AWS. While it is possible
to create these identity entities outside of eksctl and provide them
as configuration options, this adds an additional burden on the end-user,
which it is preferable to avoid.

Support for providing a permissions boundary to the identity-based entities
has been added to eksctl by extending the configuration file primarily.
This decision was made based on the fact that serviceRoleARN for a cluster
is provided in a similar way today. The assumption here is that providing
a permissions boundary is not the common case, and should therefore not
pollute the CLI more than necessary.

Issue #1221