Fix AMI selection for CF template in Managed Node Groups
The current implementation is trying to determine the AMI for managed
node groups based on the AMI family value. This is incorrect because
utils.IsGPUInstanceType is expecting an instanceType value.

Therefore, when trying to create GPU managed node groups, it isn't
actually selecting the correct GPU AMI, and the creation is failing.
(Nodegroups fail to stabilize from an Internal Failure)

This is the current CF template being generated when the instance type
is `g4dn.4xlarge`:

```
"ManagedNodeGroup": {
      "Type": "AWS::EKS::Nodegroup",
      "Properties": {
        "ClusterName": "XXXX",
        "NodegroupName": "gpu-1",
        "ScalingConfig": {
          "MinSize": 1,
          "MaxSize": 2,
          "DesiredSize": 1
        },
        "Subnets": {
          "Fn::Split": [
            ",",
            {
              "Fn::ImportValue": "eksctl-XXXX-cluster::SubnetsPublic"
            }
          ]
        },
        "InstanceTypes": [
          "g4dn.4xlarge"
        ],
        "AmiType": "AL2_x86_64",
        "RemoteAccess": {
          "Ec2SshKey": "XXXX"
        },
        "NodeRole": {
          "Fn::GetAtt": "NodeInstanceRole.Arn"
        },
        "Labels": {
          "alpha.eksctl.io/cluster-name": "XXXX",
          "alpha.eksctl.io/nodegroup-name": "XXXX"
        },
        "Tags": {
          "alpha.eksctl.io/nodegroup-name": "XXXX",
          "alpha.eksctl.io/nodegroup-type": "managed"
        }
      }
    },
```

`AL2_x86_64` should be `AL2_x86_64_GPU`