Allow new resources to be created during cluster stack update
ClusterResourceSet builds a CloudFormation template using the ClusterConfig as input and skips creation of certain resources if the user has supplied their own resource (e.g., ClusterConfig.VPC.SecurityGroup) This works fine during cluster creation.

During an update, the VPC fields in ClusterConfig supplied by the user are ignored because user-supplied updates to VPC are not (yet) supported. However, before building a new template, the code loads the VPC config from the existing cluster's stack output, which results in ClusterResourceSet skipping execution of code that'd otherwise be executed when these fields aren't supplied by the user. As a consequence, any update logic that adds code that's conditionally executed, depending on whether a field was user-supplied or not, will be skipped and the updated stack will not have the new resources.

This change fixes creating resources that are created when VPC.SecurityGroup and VPC.SharedNodeSecurityGroup are not user-supplied.