Fix upgrading managed nodegroups
Currently, AMIs published for EKS worker nodes are not immediately available for use with managed nodegroups. To work around this, this change uses the AWS EKS Nodegroup API and leaves the ReleaseVersion unset to allow the default behaviour of the API choosing the latest available AMI version.

This change also switches from updating the CloudFormation stack to using the UpdateNodegroupVersion API call, because with the ReleaseVersion no longer being set to the latest AMI release obtained via SSM, some other field must be updated in order to trigger a stack update. Rather than use a hack like updating a label on the nodegroup, we simply use the UpdateNodegroupVersion API call which also gives more control over the update process.