Fix upgrading nodegroups with old template format (#2572)
* Fix upgrading nodegroups with old template format

goformation unmarshals an attribute reference as an array. For old templates, this was causing the reference to `"Fn::GetAtt": "NodeInstanceRole.Arn"` to be unmarshalled as `"Fn::GetAtt": ["NodeInstanceRole", "Arn"]`, which, in turn, was triggering an update on this resource. However, when upgrading a nodegroup, CloudFormation does not permit combining release version updates with other updates, and this resulted in an error.

* Use LT instead of Compare