Undo setting cgroup drivers to systemd
Since we have stopped providing custom bootstrapping scripts and are now
relying on what comes with the AMI, finding the right places to change
the setting for both kubelet and docker is trickier. GPU nodes now fail
to come up. The GPU AMIs are going through some changes, so those
locations will be even harder to pin down for a while.
I am undoing this for now, and plan to look into it later when I have
more time.