Add support for Karpenter defaultInstanceProfile setting (#4751)
* Update generated files

* Add support for Karpenter defaultInstanceProfile setting

Added support to pass in `defaultInstanceProfile` parameter for Karpenter which configures the default IAM instance profile via the Karpenter Helm chart. When enabled this configures Karpenter to use the node IAM instance profile created by CloudFormation as the default instance profile, which can be overridden in Provisioners if required.

* Update generated files

* WIP commit

* Add support for Karpenter defaultInstanceProfile setting

Added support to pass in `defaultInstanceProfile` parameter for Karpenter which configures the default IAM instance profile via the Karpenter Helm chart. When enabled this configures Karpenter to use the node IAM instance profile created by CloudFormation as the default instance profile, which can be overridden in Provisioners if required.

* Update generated files

* Add support for Karpenter defaultInstanceProfile setting

Added support to pass in `defaultInstanceProfile` parameter for Karpenter which configures the name of the default IAM instance profile created by the Karpenter integration. This change also automatically configures the default IAM instance profile via the Karpenter Helm chart.

* Add support for Karpenter defaultInstanceProfile setting

Added support to pass in `defaultInstanceProfile` parameter for Karpenter which configures the name of the default IAM instance profile created by the Karpenter integration. This change also automatically configures the default IAM instance profile via the Karpenter Helm chart.

* Add support for Karpenter defaultInstanceProfile setting

Added support to pass in `defaultInstanceProfile` parameter for Karpenter which configures the default IAM instance profile via the Karpenter Helm chart. When enabled this configures Karpenter to use the node IAM instance profile created by CloudFormation as the default instance profile, which can be overridden in Provisioners if required.

Co-authored-by: Gergely Brautigam <182850+Skarlso@users.noreply.github.com>