Differentiate among accelerated hardware providers
`IsGPUInstanceType` groups all together NVIDIA and inferentia instances
which makes it difficult to reuse for OS'es that don't vend one image
with support for multiple accelerated hardware.

With this change, the instances with support for NVIDIA GPUs are
separated in its own function (like inferentia instances), which is
called by `IsGPUInstanceType`.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>