Skip NVIDIA device plugin daemonset in Bottlerocket node groups
The Bottlerocket NVIDIA AMI already provides the NVIDIA device plugin,
so Bottlerocket node groups don't require the NVIDIA device plugin
daemonset.

This change makes sure that clusters with only Bottlerocket nodegroups
don't get the NVIDIA device plugin daemonset. If any of the nodegroups
in the cluster's configuration uses an OS that requires the device
plugin, it will be installed.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>